Update code so there is no more global File object for the data file
Now, $data_file is a string (expanded path to data file), not a File
object. Instead of trying to use a global File object, tracking will now
reopen the data file whenever it's needed.

This makes using File and CSV in back compatible ways (for Ruby 1.8)
easier. In fact, it should/might fix some issues with tracking breaking
in Ruby 1.8 because of missing features in File and CSV.